Revert "Partial implementation of correlation"
This reverts commit 2b9749e7163273e3ccd58501575241b1660f5425.

+++ cpp/commander/LearningCommander.hpp
@@ -28,10 +28,10 @@ class LearningCommander : public StreamObserver{
 		Sphero *sphero;
 		Target *target;
 		
-		double inputMins[INPUTSIZE];
-		double inputMaxs[INPUTSIZE];
-		double outputMins[OUTPUTSIZE];
-		double outputMaxs[OUTPUTSIZE];
+		float inputMins[INPUTSIZE];
+		float inputMaxs[INPUTSIZE];
+		float outputMins[OUTPUTSIZE];
+		float outputMaxs[OUTPUTSIZE];
 		
 		double normalize(double value, double min, double max);
 		double denormalize(double value, double min, double max);
